v1.0 Load from APIs
This project is from the Platzi course: "jQuery to Javascript" where @LeonidasEsteban teaches how to use Vanilla JS instead of jQuery.
This project has some animations, some basic HTML, but the main goal is to learn Javascript.
The data for this project is from public APIs (randomuser.me for friends list, yts.mx for movies) and also has cache loading and saving with localStorage.